fix: ディレクトリ構成を変更
