+++ src/layouts/History/Fridge/FridgeView.module.css
