+++ src/layouts/History/Fridge/IceCell/IceCell.module.css
